requirements.txt
numpy==1.21.2
pandas==1.2.3
PyYAML==6.0.1
scikit_learn==1.3.1
torch==1.13.1
